Clean up integration tests.
- Remove all UI tests
