+++ plugin/src/test/java/io/jenkins/plugins/analysis/warnings/StepsITest.java
@@ -29,12 +29,16 @@ import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
 import hudson.model.Result;
 import hudson.model.Run;
+import hudson.model.Slave;
 import hudson.model.UnprotectedRootAction;
 import hudson.util.HttpResponses;
 
 import io.jenkins.plugins.analysis.core.model.AnalysisResult;
+import io.jenkins.plugins.analysis.core.model.FileNameRenderer;
+import io.jenkins.plugins.analysis.core.model.IssuesDetail;
 import io.jenkins.plugins.analysis.core.model.ReportScanningTool;
 import io.jenkins.plugins.analysis.core.model.ResultAction;
+import io.jenkins.plugins.analysis.core.model.SourceDetail;
 import io.jenkins.plugins.analysis.core.model.Tool;
 import io.jenkins.plugins.analysis.core.steps.PublishIssuesStep;
 import io.jenkins.plugins.analysis.core.steps.ScanForIssuesStep;
@@ -58,6 +62,40 @@ import static io.jenkins.plugins.analysis.core.assertions.Assertions.*;
 @SuppressWarnings({"PMD.ExcessiveImports", "PMD.ExcessiveClassLength", "checkstyle:ClassDataAbstractionCoupling", "checkstyle:ClassFanOutComplexity"})
 public class StepsITest extends IntegrationTestWithJenkinsPerTest {
     private static final String NO_QUALITY_GATE = "";
+    private static final String JAVA_CONTENT = "public class Test {}";
+
+    /**
+     * Verifies that affected source files are copied to Jenkins build folder, even if the master - agent security
+     * is active, see JENKINS-56007 for details.
+     */
+    @Test @org.jvnet.hudson.test.Issue("JENKINS-56007")
+    public void shouldCopySourcesIfMasterAgentSecurityIsActive() {
+        Slave agent = createAgentWithEnabledSecurity("agent");
+
+        WorkflowJob project = createPipeline();
+
+        createFileInAgentWorkspace(agent, project, "Test.java", JAVA_CONTENT);
+
+        project.setDefinition(new CpsFlowDefinition("node('agent') {\n"
+                + "    echo '[javac] Test.java:39: warning: Test Warning'\n"
+                + "    recordIssues tool: java()\n"
+                + "}", true));
+
+        AnalysisResult result = scheduleSuccessfulBuild(project);
+        assertThat(result).hasNoErrorMessages();
+        assertThat(result).hasTotalSize(1);
+        assertThat(getConsoleLog(result)).contains("1 copied", "0 not in workspace", "0 not-found", "0 with I/O error");
+
+        // TODO: check for the links in the table model
+        assertThat(getSourceCode(result, 0)).contains(JAVA_CONTENT);
+    }
+
+    private String getSourceCode(final AnalysisResult result, final int rowIndex) {
+        IssuesDetail target = (IssuesDetail) result.getOwner().getAction(ResultAction.class).getTarget();
+        String sourceCodeUrl = new FileNameRenderer(result.getOwner()).getSourceCodeUrl(result.getIssues().get(rowIndex));
+        SourceDetail dynamic = (SourceDetail) target.getDynamic(sourceCodeUrl.replaceAll("/#.*", ""), null, null);
+        return dynamic.getSourceCode();
+    }
 
     /**
      * Runs a pipeline and verifies the {@code scanForIssues} step has some whitelisted methods.
